refactor zarr and n5 with common superclasses, and refactor datareader in simpler composition instead of inheritance
