+++ util/src/main/scala/com/scalableminds/util/enumeration/ExtendedEnumeration.scala
@@ -4,7 +4,6 @@ import play.api.libs.json.{Format, Json}
 
 abstract class ExtendedEnumeration extends Enumeration {
   implicit val format: Format[Value] = Json.formatEnum(this)
-  def fromString(s: String): Option[Value] = {
+  def fromString(s: String): Option[Value] =
     values.find(_.toString == s)
-  }
 }
