+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/dataformats/BucketProvider.scala
@@ -1,12 +1,14 @@
 package com.scalableminds.webknossos.datastore.dataformats
 
 import com.scalableminds.util.tools.{Fox, FoxImplicits}
+import com.scalableminds.webknossos.datastore.dataformats.zarr.RemoteSourceDescriptor
 import com.scalableminds.webknossos.datastore.models.BucketPosition
 import com.scalableminds.webknossos.datastore.models.requests.DataReadInstruction
-import com.scalableminds.webknossos.datastore.storage.DataCubeCache
+import com.scalableminds.webknossos.datastore.storage.{DataCubeCache, FileSystemsHolder}
 import com.typesafe.scalalogging.LazyLogging
 import net.liftweb.common.{Box, Empty}
 
+import java.nio.file.{FileSystem, Path}
 import scala.concurrent.ExecutionContext
 
 trait BucketProvider extends FoxImplicits with LazyLogging {
@@ -37,5 +39,21 @@ trait BucketProvider extends FoxImplicits with LazyLogging {
 
   def bucketStream(version: Option[Long] = None): Iterator[(BucketPosition, Array[Byte])] =
     Iterator.empty
+
+  protected def remotePathFrom(remoteSource: RemoteSourceDescriptor): Option[Path] =
+    FileSystemsHolder.getOrCreate(remoteSource).map { fileSystem: FileSystem =>
+      fileSystem.getPath(remoteSource.remotePath)
+    }
+
+  protected def localPathFrom(readInstruction: DataReadInstruction, relativeMagPath: String): Option[Path] = {
+    val magPath = readInstruction.baseDir
+      .resolve(readInstruction.dataSource.id.team)
+      .resolve(readInstruction.dataSource.id.name)
+      .resolve(readInstruction.dataLayer.name)
+      .resolve(relativeMagPath)
+    if (magPath.toFile.exists()) {
+      Some(magPath)
+    } else None
+  }
 
 }
