+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/datareaders/FileSystemStore.scala
@@ -1,8 +1,8 @@
-package com.scalableminds.webknossos.datastore.jzarr
+package com.scalableminds.webknossos.datastore.datareaders
+
+import net.liftweb.util.Helpers.tryo
 
 import java.nio.file.{Files, Path}
-
-import net.liftweb.common.Box.tryo
 
 class FileSystemStore(val internalRoot: Path) {
   def readBytes(key: String): Option[Array[Byte]] = {
