+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/datareaders/jzarr/LICENSE
