+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/datareaders/jzarr/ZarrArray.scala
@@ -1,0 +1,55 @@
+package com.scalableminds.webknossos.datastore.datareaders.jzarr
+
+import java.io.IOException
+import java.nio.charset.StandardCharsets
+import java.nio.file.Path
+import com.scalableminds.webknossos.datastore.datareaders.{
+  ArrayOrder,
+  AxisOrder,
+  ChunkReader,
+  ChunkUtils,
+  DatasetArray,
+  DatasetHeader,
+  FileSystemStore,
+  MultiArrayUtils,
+  DatasetPath
+}
+import com.typesafe.scalalogging.LazyLogging
+import play.api.libs.json.{JsError, JsSuccess, Json}
+
+object ZarrArray extends LazyLogging {
+  private val chunkSizeLimitBytes = 64 * 1024 * 1024
+
+  @throws[IOException]
+  def open(path: Path, axisOrderOpt: Option[AxisOrder]): ZarrArray = {
+    val store = new FileSystemStore(path)
+    val rootPath = new DatasetPath("")
+    val headerPath = rootPath.resolve(ZarrHeader.FILENAME_DOT_ZARRAY)
+    val headerBytes = store.readBytes(headerPath.storeKey)
+    if (headerBytes.isEmpty)
+      throw new IOException(
+        "'" + ZarrHeader.FILENAME_DOT_ZARRAY + "' expected but is not readable or missing in store.")
+    val headerString = new String(headerBytes.get, StandardCharsets.UTF_8)
+    val header: ZarrHeader =
+      Json.parse(headerString).validate[ZarrHeader] match {
+        case JsSuccess(parsedHeader, _) =>
+          parsedHeader
+        case errors: JsError =>
+          throw new Exception("Validating json as zarr header failed: " + JsError.toJson(errors).toString())
+      }
+    if (header.bytesPerChunk > chunkSizeLimitBytes) {
+      throw new IllegalArgumentException(
+        f"Chunk size of this Zarr Array exceeds limit of $chunkSizeLimitBytes, got ${header.bytesPerChunk}")
+    }
+    new ZarrArray(rootPath, store, header, axisOrderOpt.getOrElse(AxisOrder.asZyxFromRank(header.rank)))
+  }
+
+}
+
+class ZarrArray(relativePath: DatasetPath, store: FileSystemStore, header: DatasetHeader, axisOrder: AxisOrder)
+    extends DatasetArray(relativePath, store, header, axisOrder)
+    with LazyLogging {
+
+  override protected val chunkReader: ChunkReader =
+    ChunkReader.create(store, header)
+}
