+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/services/DSRemoteTracingstoreClient.scala
@@ -4,7 +4,7 @@ import com.google.inject.Inject
 import com.scalableminds.util.tools.{Fox, FoxImplicits}
 import com.scalableminds.webknossos.datastore.DataStoreConfig
 import com.scalableminds.webknossos.datastore.dataformats.zarr.ZarrSegmentationLayer
-import com.scalableminds.webknossos.datastore.jzarr.{OmeNgffHeader, ZarrHeader}
+import com.scalableminds.webknossos.datastore.datareaders.jzarr.{OmeNgffHeader, ZarrHeader}
 import com.scalableminds.webknossos.datastore.rpc.RPC
 import com.typesafe.scalalogging.LazyLogging
 import play.api.inject.ApplicationLifecycle
